Begin cleaner edge direction handling
In particular, the directionality of the graph
is now forwards (the intuitive direction for NatLog)
and the search is explicitly doing a 'backwards' search
all the way until the push() function -- which is now agnostic
to the direction and treats the search as a forwards search again.

Code compiles, tests no longer seg fault, but it's not 'working' per
se yet.